Revert "[v0.2] Move Python dependencies into extensions"
