+++ .github/workflows/CI.yml
@@ -13,7 +13,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.9'
+          - '1.6'
           - '1'
 #          - 'nightly'
         os:
