+++ Project.toml
@@ -1,11 +1,7 @@
 name = "SymmetryReduceBZ"
 uuid = "49a35663-c880-4242-bebb-1ec8c0fa8046"
 authors = ["Jeremy Jorgensen <jerjorg@gmail.com>"]
-version = "0.2.0"
-
-[weakdeps]
-PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
-PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
+version = "0.1.9"
 
 [deps]
 CDDLib = "3391f64e-dcde-5f30-b752-e11513730f60"
@@ -13,6 +9,8 @@ Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
 Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Polyhedra = "67491407-f73d-577b-9b50-8179a7c68029"
+PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
+PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 QHull = "a8468747-bd6f-53ef-9e5c-744dbc5c59e7"
 
 [compat]
@@ -20,17 +18,14 @@ Aqua = "0.8"
 CDDLib = "0.9"
 Combinatorics = "1"
 Distances = "0.10"
-LinearAlgebra = "1.9"
+LinearAlgebra = "1.6"
 Polyhedra = "0.6,0.7"
 PyCall = "1.9"
 PyPlot = "2.9"
 QHull = "0.2"
 SymPy = "1,2"
-Test = "1.9"
-julia = "1.9"
-
-[extensions]
-SymmetryReduceBZPyPlotExt = ["PyPlot", "PyCall"]
+Test = "1.6"
+julia = "1.6"
 
 [extras]
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
@@ -39,7 +34,6 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 QHull = "a8468747-bd6f-53ef-9e5c-744dbc5c59e7"
 SymPy = "24249f21-da20-56a4-8eb1-6a02cf4ae2e6"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
 
 [targets]
-test = ["Aqua", "Test", "SymPy", "CDDLib", "LinearAlgebra", "QHull", "PyPlot"]
+test = ["Aqua", "Test", "SymPy", "CDDLib", "LinearAlgebra", "QHull"]
